src/nim_client.py
import requests
from typing import Dict, Any
import onnx
import onnxruntime

class NIMClient:
    def __init__(self, base_url: str):
        self.base_url = base_url

    def deploy_model(self, model_name: str, model_path: str) -> Dict[str, Any]:
        url = f"{self.base_url}/models/{model_name}"
        
        # Convert model to ONNX
        onnx_path = f"{model_path}_onnx"
        self._convert_to_onnx(model_path, onnx_path)
        
        # Optimize ONNX model
        optimized_onnx_path = f"{onnx_path}_optimized"
        self._optimize_onnx(onnx_path, optimized_onnx_path)
        
        payload = {"model_path": optimized_onnx_path}
        response = requests.post(url, json=payload)
        return response.json()

    def inference(self, model_name: str, input_data: Dict[str, Any]) -> Dict[str, Any]:
        url = f"{self.base_url}/models/{model_name}/infer"
        response = requests.post(url, json=input_data)
        return response.json()

    def get_model_status(self, model_name: str) -> Dict[str, Any]:
        url = f"{self.base_url}/models/{model_name}/status"
        response = requests.get(url)
        return response.json()

    def undeploy_model(self, model_name: str) -> Dict[str, Any]:
        url = f"{self.base_url}/models/{model_name}"
        response = requests.delete(url)
        return response.json()

    def _convert_to_onnx(self, model_path: str, onnx_path: str):
        # Implementation of PyTorch to ONNX conversion
        pass

    def _optimize_onnx(self, onnx_path: str, optimized_onnx_path: str):
        onnx_model = onnx.load(onnx_path)
        optimized_model = onnxruntime.optimize_model(
            onnx_model,
            ["eliminate_unused_initializer", "fuse_matmul_add_bias_into_gemm"]
        )
        onnx.save(optimized_model, optimized_onnx_path)